+++ app/src/main/java/fr/neon/e_kfc_v2/Model.kt
@@ -3,9 +3,9 @@ package fr.neon.e_kfc_v2
 import kotlin.properties.Delegates
 
 class Model {
-    var temperature by Delegates.notNull<Double>()
+    var temperature by Delegates.notNull<Long>()
 
-    constructor(temperature: Double) {
+    constructor(temperature: Long) {
         this.temperature = temperature
     }
